+++ Collections/Observables/ObservableSortedList.cs
@@ -227,11 +227,16 @@ namespace Utilities.DotNet.Collections.Observables
             }
 
 #if BULK_NOTIFY_RANGE_ACTIONS
+            if( count == 0 )
+            {
+                return;
+            }
+
             var removedItems = m_list.GetRange( index, count );
 
             m_list.RemoveRange( index, count );
 
-            NotifyCollectionChanged( new NotifyCollectionChangedEventArgs( NotifyCollectionChangedAction.Remove, removedItems, index ) );
+            NotifyCollectionChanged( new NotifyCollectionChangedEventArgs( NotifyCollectionChangedAction.Remove, removedItems ) );
 #else
             for( int i = 0; i < count; i++ )
             {
